Add FormAction
User can enter a straightforward case of predefined values for each
category. Define categories better, so that user has more freedom in
inputing.

So far works only for one truck. Add looping.
And unhappy path stories.

Logging is disabled. Find a function that is activated only once, for
example when a user enters a name of the company.
Find where is default logging for the whole conversation.